dafny-api/pyproject.toml
[tool.poetry]
name = "dafny-api"
version = "2.9.0"
description = "Dafny API for FM Playground"
authors = ["Soaibuzzaman <soaibuzzaman@uni-weimar.de>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.120.1"}
httpx = "^0.27.0"
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
requests = "^2.32.3"
uvicorn = "^0.30.5"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
